+++ _flake8_tergeo/checks/ast_call.py
@@ -6,6 +6,7 @@ import ast
 from typing import cast
 
 from _flake8_tergeo.ast_util import (
+    get_imports,
     get_parent,
     is_constant_node,
     is_expected_node,
@@ -76,6 +77,7 @@ def check_call(node: ast.Call) -> IssueGenerator:
     yield from _check_deprecated_decorator(node)
     yield from _check_bad_subprocess_aliases(node)
     yield from _check_typevar_usage(node)
+    yield from _os_error_with_errno(node)
 
 
 def _check_os_walk(node: ast.Call) -> IssueGenerator:
@@ -659,4 +661,30 @@ def _check_typevar_usage(node: ast.Call) -> IssueGenerator:
         column=node.col_offset,
         issue_number="127",
         message="Use the new generic syntax instead of TypeVar.",
+    )
+
+
+def _os_error_with_errno(node: ast.Call) -> IssueGenerator:
+    if not isinstance(node.func, ast.Name):
+        return
+    if stringify(node.func) != "OSError":
+        return
+    if len(node.args) == 0:
+        return
+    if not isinstance(node.args[0], ast.Attribute):
+        return
+    if not stringify(node.args[0]).startswith("errno."):
+        return
+    if "errno" not in get_imports(node):
+        return
+
+    yield Issue(
+        line=node.lineno,
+        column=node.col_offset,
+        issue_number="019",
+        message=(
+            "The constructor of OSError does not handle errno values specially but just "
+            "threads them as the error message. To set the errno attribute, set it on the OSError "
+            "instance after the call."
+        ),
     )