+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 * Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
+- Support type aliases in FTP077 and FTP104
 
 ## 25.3.25.1