+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Next version
 
 * Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
-- Support type aliases in FTP077 and FTP104
-- Support `typing.Union` in FTP077 and FTP104
+* Support type aliases in FTP077 and FTP104
+* Support `typing.Union` in FTP077 and FTP104
+* FTP019 which finds `OSError` instantiations with `errno` constants
 
 ## 25.3.25.1