+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Next version
 
-* Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
+- Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
 - Support type aliases in FTP077 and FTP104
 - Support `typing.Union` in FTP077 and FTP104
+- FTP130 which finds usage of `string.Template`
 
 ## 25.3.25.1