+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 
+## 25.5.14.0
 - Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
 - Support type aliases in FTP077 and FTP104
 - Support `typing.Union` in FTP077 and FTP104